+++ frontend/src/pages/ResultsPage.tsx
@@ -2,27 +2,10 @@ import PlayersSection from "../components/PlayerList/PlayersSection";
 import ResultsSection from "../components/ResultsList/ResultsSection";
 
 function ResultsPage() {
-  const playerArray = [
-    { name: "플레이어1", myTurn: false },
-    { name: "플레이어2", myTurn: true },
-    { name: "플레이어3", myTurn: false },
-    { name: "플레이어4", myTurn: false },
-  ];
-
-  const resultsSet = {
-    round: 2,
-    results: [
-      { name: "플레이어2", input: "Lorem Ipsum", src: "#", isLast: false },
-      { name: "플레이어1", input: "dolor sit", src: "#", isLast: false },
-      { name: "플레이어3", input: "Lorem Ipsum", src: "#", isLast: false },
-      { name: "플레이어4", input: "consectetur", src: "", isLast: true },
-    ],
-  };
-
   return (
     <div className=" h-screen w-screen flex justify-center items-center font-hs bg-[#E7F5FF] ">
-      <PlayersSection players={playerArray} />
-      <ResultsSection resultsSet={resultsSet} />
+      <PlayersSection />
+      <ResultsSection />
     </div>
   );
 }
